autogen.sh: it's not guaranteed that libtoolize, etc. exists - exit immediately
In certain cases it's possible that the libtoolize or other commands
are not exists, therefore this needs to be checked.
In this patch the trivial solution is added due to the simplicity of
the script: each command got the '|| exit 1' "suffix", if something
wents wrong, it's exited with status 1, as expected, instead of
continuing the run of the script.